moved image detail view to Image view folder
