+++ archive/views/__init__.py
@@ -1,9 +1,10 @@
 from .comment import *
 from .image import *
+from .Image.image_list import *
+from .Image.image_detail import *
 from .attachment import *
 from .note import *
 from .person import *
 from .tag import *
 from .usermanagement import *
 from .tree import *
-from .Image.image_list import *
